Don't attempt to link libstdc++fs when using Clang
On MacOS and (I think) most BSDs the default standard library is libc++, and so attempting to link with libstdc++fs does not work.

On Linux, where (AFAIK) most distros use Clang/libstdc++, this is going to mean link-time failures when using std::filesystem. In this case the user can add Link-Flags: -lstdc++fs to their toolchain file themselves as a fix.