Early iteration of config-headers.
This is a very rudimentary version. All it does is copy and rename the
template header file and place it in the build directory in a special
subdirectory that is added as an #include-path.

There is some code duplication and cleanup necessary. There are a
few "magic strings" and "magic paths" that need to be removed as well.