Add a tmp_git_repo_factory test fixture
Provides local repos for testing parts of dds involving dependencies
without requiring a dependency to exist somewhere on the internet.
Note that a dummy username and email are set for the tmp_git_repo to
handle cases where the developer did not set a global git username and
email, such as on build servers.