Add some logging, update tests
Update the tests to not check `test_uses` but check the `libs/` layout.

`test_uses` doesn't need to be checked. It's effectively a different
target than the library proper, and it can't be `use`d or `test_use`d.
Therefore, if there is no cycle in the `uses`, there is no cycle in the
`test_uses`.